ref: Move MainGameActivity to activity package
Refactored the MainGameActivity class by moving it from its previous location to the 'activity' package.

Files modified:
- AndroidManifest.xml
- MainGameActivity.java
- MainGameView.java
- MainView.java
- activity_main_game.xml
