+++ app/src/main/res/layout/activity_main_game.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".MainGameActivity"
+    tools:context=".activity.MainGameActivity"
     tools:layout_editor_absoluteX="0dp"
     tools:layout_editor_absoluteY="-5dp">
